feat: Move API call from search page to landing page
